Removed unnecessary layer of abstraction for "web" stuff.

+++ net/cpp/db/net/ConnectionWrapper.h
@@ -111,14 +111,14 @@ public:
     * 
     * @return the total number of bytes read so far.
     */
-   virtual unsigned long long getBytesRead();
+   virtual uint64_t getBytesRead();
    
    /**
     * Gets the number of bytes written to this Connection so far.
     * 
     * @return the number of bytes written so far.
     */
-   virtual unsigned long long getBytesWritten();
+   virtual uint64_t getBytesWritten();
    
    /**
     * Sets the read timeout for this Connection. This is the amount of
@@ -126,7 +126,7 @@ public:
     * 
     * @param timeout the read timeout in milliseconds (0 for no timeout).
     */
-   virtual void setReadTimeout(unsigned long timeout);
+   virtual void setReadTimeout(uint32_t timeout);
    
    /**
     * Sets the write timeout for this Connection. This is the amount of
@@ -134,7 +134,7 @@ public:
     * 
     * @param timeout the writetimeout in milliseconds (0 for no timeout).
     */
-   virtual void setWriteTimeout(unsigned long timeout);
+   virtual void setWriteTimeout(uint32_t timeout);
    
    /**
     * Marks this Connection as secure or non-secure.
